[refactor] Fix review issues for item submission: class names, interface structure, assertion methods
