+++ judgels-backends/sandalphon/sandalphon-api/src/main/java/judgels/sandalphon/api/submission/bundle/Grading.java
@@ -1,0 +1,14 @@
+package judgels.sandalphon.api.submission.bundle;
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import java.util.Optional;
+import org.immutables.value.Value;
+
+@Value.Immutable
+@JsonDeserialize(as = ImmutableGrading.class)
+public interface Grading {
+    Verdict getVerdict();
+    Optional<Integer> getScore();
+
+    class Builder extends ImmutableGrading.Builder {}
+}
