+++ judgels-backends/sandalphon/sandalphon-client/src/main/java/judgels/sandalphon/submission/bundle/FixedVerdictItemSubmissionGrader.java
@@ -3,14 +3,14 @@ package judgels.sandalphon.submission.bundle;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Optional;
 import judgels.sandalphon.api.problem.bundle.Item;
-import judgels.sandalphon.api.submission.bundle.ItemSubmission.Grading;
+import judgels.sandalphon.api.submission.bundle.Grading;
 import judgels.sandalphon.api.submission.bundle.Verdict;
 
-public class SetVerdictItemSubmissionGrader implements ItemSubmissionGrader {
+public class FixedVerdictItemSubmissionGrader implements ItemSubmissionGrader {
     private final ObjectMapper objectMapper;
     private final Verdict verdict;
 
-    public SetVerdictItemSubmissionGrader(ObjectMapper objectMapper, Verdict verdict) {
+    public FixedVerdictItemSubmissionGrader(ObjectMapper objectMapper, Verdict verdict) {
         this.objectMapper = objectMapper;
         this.verdict = verdict;
     }
