+++ judgels-backends/sandalphon/sandalphon-client/src/main/java/judgels/sandalphon/submission/bundle/ItemSubmissionGrader.java
@@ -1,7 +1,7 @@
 package judgels.sandalphon.submission.bundle;
 
 import judgels.sandalphon.api.problem.bundle.Item;
-import judgels.sandalphon.api.submission.bundle.ItemSubmission.Grading;
+import judgels.sandalphon.api.submission.bundle.Grading;
 
 public interface ItemSubmissionGrader {
     Grading grade(Item item, String answer);
