+++ judgels-backends/sandalphon/sandalphon-client/src/test/java/judgels/sandalphon/submission/bundle/FixedVerdictItemSubmissionGraderTests.java
@@ -1,16 +1,16 @@
 package judgels.sandalphon.submission.bundle;
 
-import static org.assertj.core.api.Java6Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import judgels.sandalphon.api.problem.bundle.Item;
 import judgels.sandalphon.api.problem.bundle.ItemType;
-import judgels.sandalphon.api.submission.bundle.ItemSubmission.Grading;
+import judgels.sandalphon.api.submission.bundle.Grading;
 import judgels.sandalphon.api.submission.bundle.Verdicts;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class SetVerdictItemSubmissionGraderTests {
+public class FixedVerdictItemSubmissionGraderTests {
     private ObjectMapper mapper;
     private Item item;
 
@@ -27,23 +27,21 @@ public class SetVerdictItemSubmissionGraderTests {
 
     @Test
     void pending_manual_grading() {
-        SetVerdictItemSubmissionGrader grader = new SetVerdictItemSubmissionGrader(
+        FixedVerdictItemSubmissionGrader grader = new FixedVerdictItemSubmissionGrader(
                 mapper, Verdicts.PENDING_MANUAL_GRADING);
 
         Grading grading = grader.grade(item, "any answer");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.PENDING_MANUAL_GRADING.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.PENDING_MANUAL_GRADING.getName());
-        assertThat(grading.getScore().isPresent()).isFalse();
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.PENDING_MANUAL_GRADING);
+        assertThat(grading.getScore()).isEmpty();
     }
 
     @Test
     void grading_not_needed() {
-        SetVerdictItemSubmissionGrader grader = new SetVerdictItemSubmissionGrader(
+        FixedVerdictItemSubmissionGrader grader = new FixedVerdictItemSubmissionGrader(
                 mapper, Verdicts.GRADING_NOT_NEEDED);
 
         Grading grading = grader.grade(item, "any answer");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.GRADING_NOT_NEEDED.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.GRADING_NOT_NEEDED.getName());
-        assertThat(grading.getScore().isPresent()).isFalse();
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.GRADING_NOT_NEEDED);
+        assertThat(grading.getScore()).isEmpty();
     }
 }
