+++ judgels-backends/sandalphon/sandalphon-client/src/test/java/judgels/sandalphon/submission/bundle/MultipleChoiceItemSubmissionGraderTests.java
@@ -1,12 +1,12 @@
 package judgels.sandalphon.submission.bundle;
 
-import static org.assertj.core.api.Java6Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import judgels.gabriel.api.Verdicts;
 import judgels.sandalphon.api.problem.bundle.Item;
 import judgels.sandalphon.api.problem.bundle.ItemType;
-import judgels.sandalphon.api.submission.bundle.ItemSubmission.Grading;
+import judgels.sandalphon.api.submission.bundle.Grading;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -44,10 +44,8 @@ public class MultipleChoiceItemSubmissionGraderTests {
         MultipleChoiceItemSubmissionGrader grader = new MultipleChoiceItemSubmissionGrader(mapper);
 
         Grading grading = grader.grade(testItem1, "b");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.ACCEPTED.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.ACCEPTED.getName());
-        assertThat(grading.getScore().isPresent()).isTrue();
-        assertThat(grading.getScore().get()).isEqualTo(4);
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.ACCEPTED);
+        assertThat(grading.getScore()).contains(4);
     }
 
     @Test
@@ -56,28 +54,20 @@ public class MultipleChoiceItemSubmissionGraderTests {
         Grading grading;
 
         grading = grader.grade(testItem1, "a");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.WRONG_ANSWER.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.WRONG_ANSWER.getName());
-        assertThat(grading.getScore().isPresent()).isTrue();
-        assertThat(grading.getScore().get()).isEqualTo(-1);
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.WRONG_ANSWER);
+        assertThat(grading.getScore()).contains(-1);
 
         grading = grader.grade(testItem1, "c");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.WRONG_ANSWER.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.WRONG_ANSWER.getName());
-        assertThat(grading.getScore().isPresent()).isTrue();
-        assertThat(grading.getScore().get()).isEqualTo(-1);
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.WRONG_ANSWER);
+        assertThat(grading.getScore()).contains(-1);
 
         grading = grader.grade(testItem1, "an answer which is not included in item choices");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.WRONG_ANSWER.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.WRONG_ANSWER.getName());
-        assertThat(grading.getScore().isPresent()).isTrue();
-        assertThat(grading.getScore().get()).isEqualTo(-1);
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.WRONG_ANSWER);
+        assertThat(grading.getScore()).contains(-1);
 
         grading = grader.grade(testItem1, "");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.WRONG_ANSWER.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.WRONG_ANSWER.getName());
-        assertThat(grading.getScore().isPresent()).isTrue();
-        assertThat(grading.getScore().get()).isEqualTo(-1);
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.WRONG_ANSWER);
+        assertThat(grading.getScore()).contains(-1);
     }
 
     @Test
@@ -85,8 +75,7 @@ public class MultipleChoiceItemSubmissionGraderTests {
         MultipleChoiceItemSubmissionGrader grader = new MultipleChoiceItemSubmissionGrader(mapper);
 
         Grading grading = grader.grade(testItem2, "b");
-        assertThat(grading.getVerdict().getCode()).isEqualTo(Verdicts.INTERNAL_ERROR.getCode());
-        assertThat(grading.getVerdict().getName()).isEqualTo(Verdicts.INTERNAL_ERROR.getName());
-        assertThat(grading.getScore().isPresent()).isFalse();
+        assertThat(grading.getVerdict()).isEqualTo(Verdicts.INTERNAL_ERROR);
+        assertThat(grading.getScore()).isEmpty();
     }
 }
